Initial commit
Added a part of the assistant module. The assistant NPC itself is currently useless, but the experience and reputation rate modifiers are there, the learning spells when leveling up or logging in is there.

I did decide to make this into a module instead of LUA scripts because I am having issues using the Eluna module and I'm tired of having to wait for fixes to a module before I can use it so I just make my own instead.

It's not as.. clean as the lua alternative because I can't for the life of me figure out how to open a gossip menu without actually having a NPC standing around like I can in LUA but it will serve the exact same purpose.

For some reason the SQL file is not being copied over either, another thing I can't figure out why.