Spawn point
Added the ability to set new characters spawn point to Stormwind City and Orgrimmar. There is also an option to set the same for death knight, for instance if the skip dk area module is used.

I decided not to have an option for setting the spawn point to Dalaran but might add one in the future.