Partial rewrite
Moving every part that uses items to the database. It's easier for a player to see what a certain item does when there's a tooltip.

I will be adding a 'reset'-like script that reverts any changes done to the database. Certain items will require modifying buy and sell prices and those should be easy to reset.