Changes
Began changing everything from hardcoded C++ values to rows in the SQL tables. This will let a user change what they want to and make it easier to add or remove entries without having to change the source every time. It will also be a little neater.

The next step is to do the same thing for every single spell used by the module.