Rewrite
Began a rewrite of the module to use the database to store everything rather than hard-coding it into the module. I will also be adding additional variables like allowing obtaining spells automatically but give an option to disable it for quested spells.

I am very new to C++, so the code is probably nowhere near good but it does what I want it to do so I can build on top of it and change things in the future to improve them.

This does mean that so far only the spawn points and riding skills are implemented in terms of spells. It took me hours to get this to work exactly like I want it to, and doing every class spell will take a long time since I want to do it right from the start with comments and such.