Fixes and additions
Added more checks for the mounts to support class-specific, race-specific and faction-specific mounts too.
Replaced the break with continue so it doesn't end the for loop whenever it reaches an entry it has to skip.

I went with multiple continues instead of mashing them all together to make it more readable.
Added alliance flying mounts and human race mounts.