Active zone
Added the ability to set creatures active that have players in the same zone. It'll reduce processing power since not every zone will be active unless they all have players in them.

The original plan was to have whole maps active instead of just zones but I decided against that in the end as seen in the code.

I'm sure there are many ways of doing this better but it works for the time being.