Professions
Added the ability to let players max out their professions by spending gold. Each rank can be enabled or disabled by config. This is helpful with a low population or can be used to let people rank up to say level 300 and have them level the rest themselves.

Did a lot of cleanup and restructured the modules files to what I do nowadays.