Instances & Text
Added the ability to reset heroic dungeons and raids. It needs some testing to make sure there are no issues so I can't give any guarantees on the quality yet. I have a few tweaks I want to make to slightly reduce the amount of code but I'll do that at a later date.
Replaced all the texts with defined strings. It should make it easier to translate since you don't have to browse through all of the code to do so.