Elixirs
Added elixirs to the assistant.

I will be adding buff food and enchants too but it's going to take time since there are so many of them.

Commented out the update queries that change the price of the items. I'm leaving the queries in just in case someone wants them but I rather have a small buy price for these items.